pyproject.toml
[tool.poetry]
name = "progphilbot"
version = "0.1.0"
description = ""
authors = ["Jedddy <jedbalita25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.0"
python-dotenv = "^0.21.0"
pyyaml = "^6.0"
flake8 = "^6.0.0"
asyncpg = "^0.27.0"
requests = "^2.28.2"
yoyo-migrations = "^8.2.0"
psycopg2 = "^2.9.5"
cloudscraper = "^1.2.69"
bs4 = "^0.0.1"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
